Remove unnecessary C-style casts

+++ flower/getopt-long.cc
@@ -45,7 +45,7 @@ Getopt_long::parselong ()
   assert (*optnm);
 
   char const *endopt = strchr (optnm, '=');
-  size_t searchlen = (endopt) ? (size_t) (endopt - optnm) : strlen (optnm);
+  size_t searchlen = (endopt) ? (endopt - optnm) : strlen (optnm);
 
   found_option_ = 0;
   for (int i = 0; i < table_len_; i++)
